MT2/requirements.txt
absl-py==0.15.0
apex==0.1
astunparse==1.6.3
batchgenerators==0.23
cached-property==1.5.2
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
cloudpickle==2.0.0
connected-components-3d==3.2.0
cycler==0.11.0
dask==2022.2.0
distance-map==1.0.1
easydict==1.9
efficientnet-pytorch==0.7.1
faiss==1.6.3
fastremap==1.12.2
flatbuffers==1.12
fsspec==2022.2.0
future==0.18.2
gast==0.3.3
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.4
imagecodecs==2021.11.20
imagecodecs-lite==2020.1.31
importlib-metadata==4.11.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.1
linecache2==1.0.0
llvmlite==0.39.1
lmdb==0.99
locket==0.2.1
Markdown==3.3.6
matplotlib==3.0.2
MedPy==0.4.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
monai==0.9.1
networkx==2.6.3
nibabel==3.2.2
numba==0.56.2
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib==3.2.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==0.23.4
partd==1.2.0
Pillow==9.0.1
prefetch-generator==1.0.1
protobuf==3.19.4
ptflops==0.6.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==1.2.0
pynvml==11.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.14.2
scikit-learn==0.20.1
scipy==1.1.0
SimpleITK==1.2.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.7
toolz==0.11.2
torch==1.8.0
torchaudio==0.8.0
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.9.0
tqdm==4.26.0
traceback2==1.4.0
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.26.12
Werkzeug==2.0.3
wrapt==1.12.1
yacs==0.1.8
zipp==3.7.0
